1. CoTask 在await_suspend时候默认直接唤醒直到遇到CoWaiting时挂起

+++ kernel/include/kernel/comp/Utils/SortUtil.h
@@ -70,7 +70,7 @@ requires requires(T t1, T t2)
 }
 ALWAYS_INLINE void SortUtil::QuickSort(std::vector<T> &arr)
 {
-    _QuickSort(arr, 0, arr.size() - 1);
+    _QuickSort(arr, 0, static_cast<Int32>(arr.size()) - 1);
 }
 
 template<typename T>
